feat: clk-domain direct measure + PowerModes/dynamic-boost-status + clippy cleanups
- hi: Gpu::clk_domain_freq_direct (green-curve MEASURE 0x527FC458, one-call freq_khz)
- src: PowerModes capability/get/set + dynamic_boost_status (PCF 0xc80068a1)
- sys: PowerMonitor GetStatus V4/V5 sibling 0x2AD3DBAB registered (documented, no new data)
- fmt + clippy --all-targets -D warnings fully clean (Default impls, struct-literal
  init, partial_cmp rewrites, doc-list indents, example probe fixes)

+++ examples/probe_allclocks_v3.rs
@@ -25,7 +25,7 @@ const GET_ALL_CLOCK_FREQ: u32 = 0xDCB616C3; // modern
 
 fn main() {
     let _ = initialize();
-    let mut handles = [NvPhysicalGpuHandle::default(); NVAPI_MAX_PHYSICAL_GPUS as usize];
+    let mut handles = [NvPhysicalGpuHandle::default(); NVAPI_MAX_PHYSICAL_GPUS];
     let mut count = 0u32;
     unsafe { NvAPI_EnumPhysicalGPUs(&mut handles, &mut count) };
     let gpu = handles[0];
