+++ examples/probe_vfp_v1v3_diff.rs
@@ -35,7 +35,7 @@ const S3: usize = 348;
 
 fn main() {
     let _ = initialize();
-    let mut handles = [NvPhysicalGpuHandle::default(); NVAPI_MAX_PHYSICAL_GPUS as usize];
+    let mut handles = [NvPhysicalGpuHandle::default(); NVAPI_MAX_PHYSICAL_GPUS];
     let mut count = 0u32;
     unsafe { NvAPI_EnumPhysicalGPUs(&mut handles, &mut count) };
     let gpu = handles[0];
